gem work on windows

+++ templates/demo.rb
@@ -1,9 +1,17 @@
 # Remove the file on both *unix and Windows
-run "rm public/index.html"
-run "del public\\index.html"
+if Gem.win_platform?
+  run "del public\\index.html"
+else
+  run "rm public/index.html"
+end
 
 gem "browsercms"
-rake "db:create"
+if Gem.win_platform?
+  puts "        rake  db:create"
+  `rake.cmd db:create`
+else
+  rake "db:create"
+end
 route "map.routes_for_browser_cms"
 generate :browser_cms
 environment 'SITE_DOMAIN="localhost:3000"', :env => "development"
@@ -12,4 +20,9 @@ environment 'SITE_DOMAIN="localhost:3000"', :env => "production"
 environment 'config.action_view.cache_template_loading = false', :env => "production"
 environment 'config.action_controller.page_cache_directory = RAILS_ROOT + "/public/cache/"', :env => "production"
 generate :browser_cms_demo_site
-rake "db:migrate"
+if Gem.win_platform?
+  puts "        rake  db:migrate"
+  `rake.cmd db:migrate`
+else
+  rake "db:migrate"
+end
